production/requirements.txt
# L9 Dependencies

# ============================================================
# CORE FRAMEWORK
# ============================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# ============================================================
# DATABASE & STORAGE
# ============================================================
neo4j>=5.14.0
redis>=5.0.0
supabase>=2.0.0

# ============================================================
# HTTP & NETWORKING
# ============================================================
requests>=2.31.0
httpx>=0.25.0

# ============================================================
# DATA & SERIALIZATION
# ============================================================
pyyaml>=6.0
orjson>=3.9.0

# ============================================================
# UTILITIES
# ============================================================
psutil>=5.9.0
python-dateutil>=2.8.2

# ============================================================
# TESTING
# ============================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# ============================================================
# MONITORING & OBSERVABILITY
# ============================================================
prometheus-client>=0.18.0

# ============================================================
# OPTIONAL (Advanced Features)
# ============================================================
# Uncomment if using advanced reasoning features
# torch>=2.0.0
# transformers>=4.35.0
# dowhy>=0.11.0
# tensorflow-probability>=0.22.0

